ci 升级openresty1.21.4.3，

+++ build-apisix-base.sh
@@ -40,17 +40,15 @@ if ([ $# -gt 0 ] && [ "$1" == "latest" ]) || [ "$version" == "latest" ]; then
     wasm_nginx_module_ver="main"
     lua_var_nginx_module_ver="master"
     grpc_client_nginx_module_ver="main"
-    amesh_ver="main"
     debug_args="--with-debug"
     OR_PREFIX=${OR_PREFIX:="/usr/local/openresty-debug"}
 else
     ngx_multi_upstream_module_ver="1.1.1"
     mod_dubbo_ver="1.0.2"
-    apisix_nginx_module_ver="1.12.0"
-    wasm_nginx_module_ver="0.6.4"
+    apisix_nginx_module_ver="1.14.0"
+    wasm_nginx_module_ver="0.6.5"
     lua_var_nginx_module_ver="v0.5.3"
-    grpc_client_nginx_module_ver="v0.4.2"
-    amesh_ver="main"
+    grpc_client_nginx_module_ver="v0.4.4"
     debug_args=${debug_args:-}
     OR_PREFIX=${OR_PREFIX:="/usr/local/openresty"}
 fi
@@ -60,7 +58,7 @@ repo=$(basename "$prev_workdir")
 workdir=$(mktemp -d)
 cd "$workdir" || exit 1
 
-or_ver="1.21.4.1"
+or_ver="1.21.4.3"
 wget --no-check-certificate https://openresty.org/download/openresty-${or_ver}.tar.gz
 tar -zxvpf openresty-${or_ver}.tar.gz > /dev/null
 
@@ -112,14 +110,6 @@ else
         grpc-client-nginx-module-${grpc_client_nginx_module_ver}
 fi
 
-if [ "$repo" == amesh ]; then
-    cp -r "$prev_workdir" ./amesh-${amesh_ver}
-else
-    git clone --depth=1 -b $amesh_ver \
-        https://github.com/api7/amesh \
-        amesh-${amesh_ver}
-fi
-
 cd ngx_multi_upstream_module-${ngx_multi_upstream_module_ver} || exit 1
 ./patch.sh ../openresty-${or_ver}
 cd ..
@@ -141,12 +131,15 @@ no_pool_patch=${no_pool_patch:-}
 grpc_engine_path="-DNGX_GRPC_CLI_ENGINE_PATH=$OR_PREFIX/libgrpc_engine.so -DNGX_HTTP_GRPC_CLI_ENGINE_PATH=$OR_PREFIX/libgrpc_engine.so"
 
 cd openresty-${or_ver} || exit 1
-# FIXME: remove this once 1.21.4.2 is released
-rm -rf bundle/LuaJIT-2.1-20220411
-lj_ver=2.1-20230119
-wget "https://github.com/openresty/luajit2/archive/v$lj_ver.tar.gz" -O "LuaJIT-$lj_ver.tar.gz"
-tar -xzf LuaJIT-$lj_ver.tar.gz
-mv luajit2-* bundle/LuaJIT-2.1-20220411
+
+if [[ "$or_ver" == 1.21.4.1 ]] || [[ "$or_ver" == 1.19.* ]]; then
+  # FIXME: remove this once 1.21.4.2 is released
+  rm -rf bundle/LuaJIT-2.1-20220411
+  lj_ver=2.1-20230119
+  wget "https://github.com/openresty/luajit2/archive/v$lj_ver.tar.gz" -O "LuaJIT-$lj_ver.tar.gz"
+  tar -xzf LuaJIT-$lj_ver.tar.gz
+  mv luajit2-* bundle/LuaJIT-2.1-20220411
+fi
 
 or_limit_ver=0.08
 if [ ! -d "bundle/lua-resty-limit-traffic-$or_limit_ver" ]; then
@@ -219,13 +212,9 @@ cd grpc-client-nginx-module-${grpc_client_nginx_module_ver} || exit 1
 sudo OPENRESTY_PREFIX="$OR_PREFIX" make install
 cd ..
 
-cd amesh-${amesh_ver} || exit 1
-sudo OPENRESTY_PREFIX="$OR_PREFIX" sh -c 'PATH="${PATH}:/usr/local/go/bin" make install'
-cd ..
-
 # package etcdctl
 ETCD_ARCH="amd64"
-ETCD_VERSION=${ETCD_VERSION:-'3.5.4'}
+ETCD_VERSION=${ETCD_VERSION:-'3.5.6'}
 ARCH=${ARCH:-$(uname -m | tr '[:upper:]' '[:lower:]')}
 
 if [[ $ARCH == "arm64" ]] || [[ $ARCH == "aarch64" ]]; then
